add docstrings and move helper functions into separate file
